+++ packages/gallery/src/components/gallery/presets/sliderGallery.js
@@ -13,14 +13,14 @@ export const fixedStyles = {
   oneRow: true,
   scrollDirection: SCROLL_DIRECTION.HORIZONTAL,
   isVertical: false,
+  groupSize: 1,
+  groupTypes: '1',
   cubeRatio: 16 / 9,
   
   //this params were moved from the presets in layoutHelper and were not tested and checked yet.
   showArrows: true,
   smartCrop: false,
   galleryType: 'Strips',
-  groupSize: 1,
-  groupTypes: '1',
   gallerySize: () => dimensionsHelper.getGalleryHeight(),
   hasThumbnails: false,
   enableScroll: true,