+++ packages/gallery/src/components/item/itemView.js
@@ -53,6 +53,8 @@ class ItemView extends GalleryComponent {
 
   init() {
     this.onItemClick = this.onItemClick.bind(this);
+    this.onItemWrapperClick = this.onItemWrapperClick.bind(this);
+    this.onItemInfoClick = this.onItemInfoClick.bind(this);
     this.onKeyPress = this.onKeyPress.bind(this);
     this.toggleFullscreenIfNeeded = this.toggleFullscreenIfNeeded.bind(this);
     this.handleItemMouseDown = this.handleItemMouseDown.bind(this);
@@ -126,8 +128,6 @@ class ItemView extends GalleryComponent {
     if (!utils.isMobile()) {
       this.props.actions.eventsListener(EVENTS.HOVER_SET, this.props.idx);
     }
-    this.onMouseOverEvent.itemIdx = this.props.idx; //VIDEOREWORK why do we need this?
-    window.dispatchEvent(this.onMouseOverEvent);
   }
 
   onKeyPress(e) {
@@ -136,6 +136,8 @@ class ItemView extends GalleryComponent {
       case 13: //enter
         e.preventDefault();
         e.stopPropagation();
+        const clickTarget = 'item-container';
+        this.props.actions.eventsListener(EVENTS.ITEM_CLICKED, {...this.props, clickTarget});
         if (this.shouldUseDirectLink()) {
           this.itemAnchor.click(); // when directLink, we want to simulate the 'enter' or 'space' press on an <a> element
         } else {
@@ -155,8 +157,21 @@ class ItemView extends GalleryComponent {
     this.props.actions.eventsListener(EVENTS.ITEM_ACTION_TRIGGERED, this.props);
   }
 
+
+
+  onItemWrapperClick(e) {
+    const clickTarget = 'item-media';
+    this.props.actions.eventsListener(EVENTS.ITEM_CLICKED, {...this.props, clickTarget});
+    this.onItemClick(e);
+  }
+
+  onItemInfoClick(e) {
+    const clickTarget = 'item-info';
+    this.props.actions.eventsListener(EVENTS.ITEM_CLICKED, {...this.props, clickTarget});
+    this.onItemClick(e);
+  }
+
   onItemClick(e) {
-    this.props.actions.eventsListener(EVENTS.ITEM_CLICKED, this.props);
     if (this.shouldUseDirectLink()) {
       return (() => { });
     }
@@ -390,7 +405,7 @@ class ItemView extends GalleryComponent {
     return dimensions;
   }
 
-  getItemTextsDetails(additionalHeight = 0) {
+  getItemTextsDetails(externalTotalInfoHeight = 0) {
     const props = utils.pick(this.props, [
       'title',
       'description',
@@ -415,7 +430,7 @@ class ItemView extends GalleryComponent {
         isSmallItem={this.isSmallItem()}
         isNarrow={this.isNarrow()}
         shouldShowButton={shouldShowButton}
-        additionalHeight={additionalHeight}
+        externalTotalInfoHeight={externalTotalInfoHeight}
         actions={{
           eventsListener: this.props.actions.eventsListener,
         }}
@@ -545,6 +560,7 @@ class ItemView extends GalleryComponent {
         key={'video' + this.props.idx}
         hover={itemHover}
         imageDimensions={imageDimensions}
+        hasLink={this.itemHasLink()}
         loadingStatus={{
           failed: this.state.failed,
           loaded: this.state.loaded,
@@ -755,19 +771,19 @@ class ItemView extends GalleryComponent {
 
     //if there is no url for videos and images, we will not render the itemWrapper
     //but will render the info element if exists, with the whole size of the item
-    const additionalHeight = this.hasRequiredMediaUrl ? 0 : style.height;
-    const additionalWidth = this.hasRequiredMediaUrl ? 0 : style.width;
+    const infoHeight = styleParams.textBoxHeight + (this.hasRequiredMediaUrl ? 0 : style.height);
+    const infoWidth = style.infoWidth + (this.hasRequiredMediaUrl ? 0 : style.width);
 
     const itemTexts = customInfoRenderer
       ? customInfoRenderer(this.props)
-      : this.getItemTextsDetails(additionalHeight);
+      : this.getItemTextsDetails(infoHeight);
 
     //TODO: move the creation of the functions that are passed to onMouseOver and onMouseOut outside
     if (itemTexts) {
       info = (
         <div style={getOuterInfoStyle(styleParams)}>
           <div
-            style={getInnerInfoStyle(styleParams, additionalHeight, additionalWidth)}
+            style={getInnerInfoStyle(styleParams, infoHeight, infoWidth)}
             className={'gallery-item-common-info ' + elementName}
             onMouseOver={() => {
               !utils.isMobile() && this.props.actions.eventsListener(
@@ -779,6 +795,7 @@ class ItemView extends GalleryComponent {
             onMouseOut={() => {
               !utils.isMobile() && this.props.actions.eventsListener(EVENTS.HOVER_SET, -1);
             }}
+            onClick={this.onItemInfoClick}
           >
             {itemTexts}
           </div>
@@ -803,10 +820,16 @@ class ItemView extends GalleryComponent {
     );
   }
 
-  getItemContainerStyles() {
-    const { styleParams, linkData, linkUrl } = this.props;
-    const containerStyleByStyleParams = getContainerStyle(styleParams);
+  itemHasLink(){
+    const { linkData, linkUrl } = this.props;
     const itemDoesntHaveLink = linkData.type === undefined && (linkUrl === undefined || linkUrl === ''); //when itemClick is 'link' but no link was added to this specific item
+    return !itemDoesntHaveLink;
+  }
+
+  getItemContainerStyles() {
+    const { styleParams } = this.props;
+    const containerStyleByStyleParams = getContainerStyle(styleParams);
+    const itemDoesntHaveLink = !this.itemHasLink(); //when itemClick is 'link' but no link was added to this specific item
     const itemStyles = {
       overflowY: styleParams.isSlideshow ? 'visible' : 'hidden',
       position: 'absolute',
@@ -1004,9 +1027,6 @@ class ItemView extends GalleryComponent {
   //-----------------------------------------| REACT |--------------------------------------------//
 
   componentDidMount() {
-    this.onMouseOverEvent = window.document.createEvent('CustomEvent'); // MUST be 'CustomEvent'
-    this.onMouseOverEvent.initCustomEvent('on_mouse_over', false, false, null);
-
     if (utils.isMobile()) {
       try {
         React.initializeTouchEvents(true);
@@ -1074,7 +1094,7 @@ class ItemView extends GalleryComponent {
   } else if (styleParams.itemClick === CLICK_ACTIONS.FULLSCREEN || styleParams.itemClick === CLICK_ACTIONS.EXPAND){
     // place share link as the navigation item
     const url = directShareLink;
-    const shouldUseDirectShareLink = false; //isSEO && !!url;
+    const shouldUseDirectShareLink = !!url;
     const linkParams = shouldUseDirectShareLink
     ? { href: url }
     : {};
@@ -1085,8 +1105,10 @@ class ItemView extends GalleryComponent {
   composeItem() {
     const { photoId, id, hash, idx, styleParams, type, url } = this.props;
 
-    //if there is no url for videos and images, we will not render the itemWrapper (but will render the info element if exists, with the whole size of the item)
-    this.hasRequiredMediaUrl = !(!url && (type === 'image' || type === 'picture' || type === 'video'));
+    //if (there is an url for video items and image items) OR text item (text item do not use media url)
+    this.hasRequiredMediaUrl = url || type === 'text';
+    //if titlePlacement !== SHOW_ON_HOVER and !this.hasRequiredMediaUrl, we will NOT render the itemWrapper (but will render the info element with the whole size of the item)
+    const isItemWrapperEmpty = styleParams.titlePlacement !== PLACEMENTS.SHOW_ON_HOVER && !this.hasRequiredMediaUrl;
     const innerDiv = (
       <div
         className={this.getItemContainerClass()}
@@ -1097,7 +1119,7 @@ class ItemView extends GalleryComponent {
         onMouseOut={() => {
           !utils.isMobile() && this.props.actions.eventsListener(EVENTS.HOVER_SET, -1);
         }}
-        onClick={this.onItemClick}
+        //onClick={this.onItemClick} //onItemClick will be called by onItemWrapperClick and onItemInfoClick
         onKeyDown={this.onKeyPress}
         tabIndex={this.getItemContainerTabIndex()}
         aria-label={this.getItemAriaLabel()}
@@ -1117,11 +1139,12 @@ class ItemView extends GalleryComponent {
             ...(this.props.styleParams.titlePlacement === PLACEMENTS.SHOW_ON_THE_LEFT && {float: 'right'})
           }}
         >
-          {this.hasRequiredMediaUrl && (<div
+          {!isItemWrapperEmpty && (<div
             data-hook="item-wrapper"
             className={this.getItemWrapperClass()}
             key={'item-wrapper-' + id}
             style={this.getItemWrapperStyles()}
+            onClick={this.onItemWrapperClick}
           >
             {this.getItemInner()}
           </div>)}