+++ packages/gallery/src/versionLogger.js
@@ -1,1 +1,1 @@
-try {console.debug('Pro Gallery Version: 1.7.20'); } catch (e) {}
+try {console.debug('Pro Gallery Version: 1.8.2'); } catch (e) {}